eliminate intermediate variables LV and LP.
I contend they aren't needed. This makes the resulting 'if' statements
that use LV and LP more readable, because we can see 'in situ'
precisely what is being compared, rather than having to dereference
the values of LF and LP by looking to the lines above to see what
they were assigned to.

I have introduced the redundancy of duplicating the expressions
'len(_values)' and 'len(_probabilities)'. If these expressions were
any more complex than this, then I would extract them as a new function
with a meaningful name. But because they are so simple, practically
actually atomic, I don't mind duplicating them. The performance
implications of duplicating the call to 'len' should be ignored until
we have measured that is makes any difference (which it won't -
reading this python file from disk to run it will take millions of
times longer than calling 'len'.)