Fiddle error messages to make then fit in 80 chars.
Also, note that I moved the check for different lengths to be first.
I consider this most fundamental, in that one could imagine the
other checks to be optional, e.g. we might decide that with empty
arrays we just return and empty sequence or something (I don't think
that would be wise, but one might have need for it.) Whereas with
arrays of different lengths, I don't think there is any sane way
we could proceed.

Interestingly, now that the check for lengths is first, the second
check, for zero lengths, could be simplified because we now only need
to test one of them. I considered not simplifying it, because it means
the second 'if' is not dependent on the first 'if' in order to
correctly work. (i.e. if future edits, done by another developer
who hasn't had time to think through all the nuances, changes or
moves or deletes the first 'if', then the second 'if' will again
need to check both len(_values) AND len(_probabilities). So it's
nice if the second if  contains both checks, so that it stands
by itself, removing any dependency on the behaviour or ordering
with respect to the first 'if'.

However, then I thought, if a future developer did make that mistake,
(i.e. removed or changed or moved the first 'if', without adding
both len checks to the second 'if', then the tests would catch it
and alert them to their mistake. So it's OK if I remove one of
the checks now, to simplify the code.

However however, then I noticed there are not tests for values
or probabilities being empty. So I added some.