convert 'if' checks to asserts.
The construct:
  if cond: raise AssertionError(message)
is identical to:
  assert \!cond, message
I prefer the second because it's slightly shorter and easier to read.
(by being less flexible than a general purpose 'if', the reader
instantly knows what an line starting 'assert ...' does, at even the
quickest of glances.)

I considered converting the third 'if' too, switching out the
ArithmeticError for another AssertionError, but decided against
it.