introduce 'cumulative' function.
Now that 'total' is a local variable, and hence evidently not used
elsewhere, it have become very clear that the for loop in '__init__'
is an example of the common 'accumulate' pattern. In Python3, I would
replace it with the function 'accumulate'
https://docs.python.org/dev/library/itertools.html#itertools.accumulate
However since we are in Python2, I'm going to write my own version
'cumulative' (so called because it isn't as general purpose as
accumulate, it takes no combining function argument, but assumes
addition.)

This now lets us look at the code in __init__  and consider it from
a highler level of abstraction, needing to read less code in order
to understand what's going on. Cumulative can be tested by itself
(meaning smaller, simpler unit tests) and can be re-used elsewhere
without needing to be tested again, thus reducing total volumes of
code in use.